requirements.txt
fastapi==0.103.0
pydantic==2.3.0
uvicorn==0.23.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
rich==13.5.2
python-dotenv==1.0.0